More conversions

+++ Sources/SwiftSDL2/Classes/Font.swift
@@ -31,7 +31,7 @@ public class Font {
 
     public init(_ fileName: String, size: Int, index: Int) {
         if Self.wasInit() == false { Self.initialize() }
-        self._fontPtr = TTF_OpenFontIndex(fileName, Int32(size), index)
+        self._fontPtr = TTF_OpenFontIndex(fileName, Int32(size), Int32(index))
     }
 
     deinit {
@@ -138,7 +138,7 @@ public extension Font {
 
     var faces: Int {
         get {
-            return TTF_FontFaces(self._fontPtr)
+            return Int(TTF_FontFaces(self._fontPtr))
         }
     }
